% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertHsMmGeneList.R
\name{convertHsMmGeneList}
\alias{convertHsMmGeneList}
\title{convertHsMmGeneList}
\usage{
convertHsMmGeneList(gene_list, convert_to = "human")
}
\arguments{
\item{gene_list}{A string with 1 gene or a list with multiple genes}

\item{convert_to}{A string defining the direction of conversion (default is human)}
}
\value{
A list of genes
}
\description{
Converts between human and mouse gene names
}
