% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_counts.R
\name{gene_counts}
\alias{gene_counts}
\title{gene_counts}
\usage{
gene_counts(seurat_object, gene_list, group = "ident")
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{gene_list}{A string with 1 gene or a list with multiple genes}

\item{group}{Metadata column of interest (default is active.ident)}
}
\value{
A data.frame with gene names, metadata, average expression, percent
    expression, and cell counts.
}
\description{
Creates a data.frame of cell counts for a list of genes.
}
