% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_numbers.R
\name{get_cell_numbers}
\alias{get_cell_numbers}
\title{get_cell_numbers}
\usage{
get_cell_numbers(
  seurat_object,
  columns = "idents",
  order = NULL,
  decr_order = TRUE
)
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{columns}{a list of one or more metadata columns (default is idents)}

\item{order}{the metadata column by which to order the columns, must be numerical}

\item{decr_order}{A bool defining whether sorted order is decreasing (default is TRUE)}
}
\value{
a data table containing the counts of cells per given column
}
\description{
Generates a table of cell numbers among given columns.
}
