% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dtplt_data.R
\name{get_dtplt_data}
\alias{get_dtplt_data}
\title{get_dtplt_data}
\usage{
get_dtplt_data(
  seu_obj,
  goi,
  group = "CellType",
  split = NULL,
  sep = "_",
  scale_100 = TRUE
)
}
\arguments{
\item{seu_obj}{A Seurat object}

\item{goi}{A list of genes of interest}

\item{group}{A column in the object's metadata to group the data by (x-axis)}

\item{split}{A column in the object to split the data by (legend)}

\item{sep}{A separator for the group and split columns}

\item{scale_100}{Scale the percentage data to 0-100}
}
\value{
A table with Gene, Group (Grp), Split (if applied), Average, scaled Average,
        Percentage, and scaled Percentage values.
}
\description{
Retrieves data from Seurat object for plotting multi-feature dot plots.
}
\author{
Sneha Korlakunta
}
