% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_gene_cat.R
\name{pct_gene_cat}
\alias{pct_gene_cat}
\title{pct_gene_cat}
\usage{
pct_gene_cat(gene_cat, pattern, max_pct, print_genes = FALSE)
}
\arguments{
\item{gene_cat}{String category for selected gene subset}

\item{pattern}{String for grep-ing genes}

\item{max_pct}{Threshold for max-cutoff for genes}

\item{print_genes}{Choose whether to print applicable genes, default is FALSE}
}
\value{
A sorted scatterplot with a set max threshold
}
\description{
Creates a scatter plot of genes matching a particular
   pattern and a set threshold for max percent
}
