% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cells.R
\name{remove_cells}
\alias{remove_cells}
\title{remove_cells}
\usage{
remove_cells(seurat_object)
}
\arguments{
\item{seurat_object}{A Seurat object}
}
\value{
A Seurat object with selected cells removed.
}
\description{
Allows user to manually remove cells in a Seurat object.
}
