% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_filt_seu.R
\name{save_filt_seu}
\alias{save_filt_seu}
\title{save_filt_seu}
\usage{
save_filt_seu(
  seurat_obj,
  counts_file,
  save_file_name,
  a_MIN_GENES_PER_CELL,
  a_MAX_GENES_PER_CELL,
  a_MIN_COUNTS_PER_CELL,
  a_MAX_COUNTS_PER_CELL,
  a_MAX_PCT_MITO,
  a_MAX_PCT_RIBO
)
}
\arguments{
\item{seurat_obj}{Unfiltered Seurat object}

\item{counts_file}{Unprocessed barcode-feature-matrix file}

\item{save_file_name}{String for filename of saved RDS file}

\item{a_MIN_GENES_PER_CELL}{Lower threshold for genes per cell}

\item{a_MAX_GENES_PER_CELL}{Upper threshold for genes per cell}

\item{a_MIN_COUNTS_PER_CELL}{Lower threshold for counts per cell}

\item{a_MAX_COUNTS_PER_CELL}{Upper threshold for counts per cell}

\item{a_MAX_PCT_MITO}{Upper threshold for mitochondrial gene expression}

\item{a_MAX_PCT_RIBO}{Upper threshold for ribosomal gene expression}
}
\value{
A Seurat object with filtered values
}
\description{
Creates a filtered Seurat object
}
