% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cells.R
\name{select_cells}
\alias{select_cells}
\title{select_cells}
\usage{
select_cells(seurat_object, selected_cells = list())
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{selected_cells}{An internal parameter}
}
\value{
A list of cells selected from the input Seurat object.
}
\description{
Allows user to manually select cells in a Seurat object.
}
