% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKyR-package.R
\docType{package}
\name{SKyR-package}
\alias{SKyR}
\alias{SKyR-package}
\title{SKyR: Miscellaneous Bioinformatics Functions}
\description{
Miscellaneous R functions that are useful to Sneha Korlakunta (and others). Primarily for bioinformatics and computational use.
}
\keyword{internal}
