% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_cells.R
\name{threshold_cells}
\alias{threshold_cells}
\title{threshold_cells}
\usage{
threshold_cells(title, values, min_per_cell, max_per_cell)
}
\arguments{
\item{title}{Name of dataset}

\item{values}{Values to be plotted on a log scale}

\item{min_per_cell}{Lower threshold value}

\item{max_per_cell}{Upper threshold value}
}
\value{
A log-scale scatterplot with upper/lower thresholds
}
\description{
Creates a logarithmic scatter plot of the given values
    with upper (gold) and lower (magenta) threshold lines on the y-axis
}
