% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uconn_macspectrum.R
\name{uconn_macspectrum}
\alias{uconn_macspectrum}
\title{uconn_macspectrum}
\usage{
uconn_macspectrum(mac_mtx, feature, select_hu_mo = "mou")
}
\arguments{
\item{mac_mtx}{A data.frame containing the gene expression levels of your dataset.
In the data.frame, rows are genes and columns are samples (e.g. individual single cells,
regular RNA-seq samples, qPCR samples, etc. see example below). The first column contains
Ensembl IDs of the genes, and other columns are gene expression levels of each cell or sample,
in UMI, FPKM/RPKM, TPM format or other relative expression formats (eg. qPCR results).
Currently, mouse or human Ensembl IDs are supported. Human genes in your sample dataset will be
automatically mapped to their murine homologs using MGI homology. The colname of the first column
should be "geneid", and colnames of the rest colnames should be sample names.}

\item{feature}{a character vector which contains any features of the samples (treated/control,
disease/healthy, etc.). The features should be of the same number and order as the samples in
the mac_mtx (e.g., cell1, cell2, cell3, etc.; or sample1, sample2, sample3, etc.).}

\item{select_hu_mo}{a string indicating if the samples are of human ("hum") or mouse ("mou"). default "mou"}
}
\value{
a data.frame containing sample names, features, MPI, and AMDI
}
\description{
Calculating MPI and AMDI of macrophage samples using MacSpectrum's algorithm.
}
\references{
https://macspectrum.uconn.edu/
}
\keyword{internal}
