% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vln_merge_data.R
\name{vln_merge_data}
\alias{vln_merge_data}
\title{vln_merge_data}
\usage{
vln_merge_data(seu_obj, goi, group = "CellType")
}
\arguments{
\item{seu_obj}{A Seurat object}

\item{goi}{A list of genes of interest}

\item{group}{A column in the object's metadata to group the data by (legend)}
}
\value{
A table with Gene, Group, and Expression values.
}
\description{
Retrieves data from Seurat object for plotting multi-feature violin plots
             (only tested for 2 conditions).
}
\author{
Sneha Korlakunta
}
