% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_clustering.R
\name{check_clustering}
\alias{check_clustering}
\title{check_clustering}
\usage{
check_clustering(seurat_object, clusters, dims = 1:30, slot = "pca")
}
\arguments{
\item{seurat_object}{A Seurat object}

\item{clusters}{the metadata column indicating the clusters to assess}

\item{dims}{A sequence of numbers indicating the number of dimensions (default is 1:30)}

\item{slot}{A dimensional reduction present in the Seurat object (default is "pca")}
}
\value{
a Seurat FeaturePlot of scores generated by the Silhouette method.
}
\description{
Generates a FeaturePlot of scores generated by the Silhouette method. Use
    to evaluate how well the cells fit their assigned cluster.
}
